+++ src/app/components/configuration/tenants/tenantsController.js
@@ -1,9 +1,25 @@
 'use strict';
 
 angular.module('liveopsConfigPanel')
-  .controller('TenantsController', ['$rootScope', '$scope', 'Session', 'Tenant', 'TenantUser', 'tenantTableConfig', 'UserPermissions', 'AuthService', 'Region', '$q', 'loEvents', 'Timezone', 'PermissionGroups', 'Alert', 'GlobalRegionsList', 'Integration',
-    function ($rootScope, $scope, Session, Tenant, TenantUser, tenantTableConfig, UserPermissions, AuthService, Region, $q, loEvents, Timezone, PermissionGroups, Alert, GlobalRegionsList, Integration) {
+  .controller('TenantsController', ['$window', '$http', '$rootScope', '$scope', 'Session', 'Tenant', 'TenantUser', 'tenantTableConfig', 'UserPermissions', 'AuthService', 'Region', '$q', 'loEvents', 'Timezone', 'PermissionGroups', 'Alert', 'GlobalRegionsList', 'Integration', 'Branding', '$translate', 'fileUpload',
+    function ($window, $http, $rootScope, $scope, Session, Tenant, TenantUser, tenantTableConfig, UserPermissions, AuthService, Region, $q, loEvents, Timezone, PermissionGroups, Alert, GlobalRegionsList, Integration, Branding, $translate, fileUpload) {
       var vm = this;
+      $scope.forms = {};
+
+      $scope.colorPickerOptions = {
+        // html attributes
+        placeholder: '',
+        format: 'hex',
+        restrictToFormat: false,
+        hue: true,
+        saturation: true,
+        case: 'upper',
+        swatch: true,
+        swatchPos: 'left',
+        round: false,
+        pos: 'bottom left',
+        inline: false,
+      };
 
       vm.loadTimezones = function () {
         $scope.timezones = Timezone.query();
@@ -50,6 +66,19 @@ angular.module('liveopsConfigPanel')
         return tenants;
       };
 
+      vm.loadBranding = function (tenantId) {
+        Branding.get({
+          tenantId: tenantId
+        }, function(response){
+          $scope.brandingForm = response;
+        }, function(error){
+          $scope.brandingForm = {};
+          if (error.status !== 404) {
+            console.log('Branding Styles Error:', error);
+          }
+        });
+      };
+
       vm.loadIntegrations = function(tenantId) {
         // tried using cachedQuery but it didn't work when selectedTenant was changed
         $scope.integrations = Integration.query({
@@ -67,14 +96,56 @@ angular.module('liveopsConfigPanel')
       };
 
       $scope.submit = function () {
-        return $scope.selectedTenant.save(null, null, function (err) {
+        $scope.selectedTenant.save(null, null, function (err) {
           if (err.data.error.attribute.parentId) {
             Alert.error(err.data.error.attribute.parentId);
           } else if (err.data.error.attribute.name) {
             Alert.error(err.data.error.attribute.name);
           }
         });
+
+        if ($scope.brandingForm !== {} && !angular.equals($scope.brandingForm, $scope.brandingForm.$original)) {
+          submitBranding();
+        }
       };
+
+      function submitBranding() {
+        if ($scope.brandingForm.updatelogo) {
+          uploadImg('logo');
+        } else if ($scope.brandingForm.updatefavicon) {
+          uploadImg('favicon');
+        } else {
+          updateBranding();
+        }
+      }
+
+      function uploadImg(type) {
+        fileUpload.uploadImg($scope.brandingForm[type + 'Selected'], type)
+          .then(function(){
+            $scope.brandingForm['update' + type] = false;
+            submitBranding();
+          }, function(error){
+            console.log(error);
+            return Alert.error($translate.instant('tenant.branding.images.' + type + '.fileSizeError'));
+          });
+      }
+
+      function updateBranding() {
+        // Update Branding Colors and apply if on current tenant
+        Branding.update({
+          tenantId: $scope.selectedTenant.id,
+          styles: $scope.brandingForm.styles
+        }, function(response) {
+          if (response.tenantId === Session.tenant.tenantId) {
+            Branding.set(response);
+          }
+          Alert.success($translate.instant('tenant.branding.updated'));
+        }, function(error) {
+          if (error.status !== 404) {
+            console.log('Branding Styles Error:', error);
+          }
+        });
+      }
 
       $scope.$on(loEvents.tableControls.itemCreate, function () {
         $scope.create();
@@ -117,6 +188,7 @@ angular.module('liveopsConfigPanel')
           });
 
           vm.loadIntegrations($scope.selectedTenant.id);
+          vm.loadBranding($scope.selectedTenant.id);
         }
       });
 
@@ -144,6 +216,57 @@ angular.module('liveopsConfigPanel')
       $scope.tableConfig = tenantTableConfig(function () {
         return $scope.tenants;
       });
+
+      $scope.resetDefaultBranding = function() {
+        Branding.update({
+          tenantId: $scope.selectedTenant.id,
+          styles: {},
+          logo: '',
+          favicon: ''
+        }, function(response){
+          $scope.brandingForm = {};
+          if (response.tenantId === Session.tenant.tenantId) {
+            Branding.set(response);
+          }
+          Alert.success($translate.instant('tenant.branding.resetDefault'));
+        }, function(errors){
+          console.log(errors);
+        });
+      };
+
+      $scope.fileSelected = function(element) {
+        $scope.brandingForm[element.name + 'Selected'] = element.files[0];
+        $scope.forms.detailsForm.$setDirty(true);
+
+        // On file selection cancel, clear form and preview
+        if ($scope.brandingForm[element.name + 'Selected'] === undefined) {
+          $scope.brandingForm[element.name] = '';
+          $scope.brandingForm.logoPreview = '';
+          return;
+        }
+
+        // File Checking for size and type
+        var maxFileSize = 1000000;
+        var fileTypesAllowed = ['image/png', 'image/jpg', 'image/jpeg', 'image/gif'];
+        if (element.name === 'favicon') {
+          fileTypesAllowed.push('image/vnd.microsoft.icon');
+        }
+        if ($scope.brandingForm[element.name + 'Selected'].size > maxFileSize || fileTypesAllowed.indexOf($scope.brandingForm[element.name + 'Selected'].type) === -1) {
+          element.value = '';
+          return Alert.error($translate.instant('tenant.branding.images.logo.fileSizeError'));
+        }
+
+        // Conditional for on form submit to upload img first
+        $scope.brandingForm['update' + element.name] = true;
+
+        // Preview Image
+        var reader  = new $window.FileReader();
+        reader.addEventListener('load', function() {
+          $scope.brandingForm[element.name + 'Preview'] = reader.result;
+        }, false);
+        reader.readAsDataURL($scope.brandingForm[element.name + 'Selected']);
+
+      };
 
       vm.loadTenants();
       vm.loadTimezones();