+++ src/app/index.js
@@ -23,7 +23,8 @@ angular.module('liveopsConfigPanel', [
     'agent-toolbar',
     'ui.sortable',
     'ui.codemirror',
-    'ngAnimate'
+    'ngAnimate',
+    'color.picker'
   ])
   .config(['$translateProvider', 'toastrConfig', '$qProvider', '$locationProvider',
     function($translateProvider, toastrConfig, qProvider, $locationProvider) {
@@ -46,10 +47,40 @@ angular.module('liveopsConfigPanel', [
       $locationProvider.hashPrefix('');
     }
   ])
-  .run(['queryCache', '$rootScope','$state', '$animate', function(queryCache, $rootScope, $state, $animate) {
+  .run(['queryCache', '$rootScope', '$state', '$animate', 'Branding', '$location', function(queryCache, $rootScope, $state, $animate, Branding, $location) {
+
+    // --- Mitel Temp Info ---
+    // Needed to style the login and forgot password pages
+    // We do not have access to the branding data until logged in
+    // This is temporary data until we have further api support
+    var mitelUrl = 'mitel-config';
+
+    var mockBrandingData = {
+      active: true,
+      favicon: '23f10080-faaa-11e6-b856-36051d50f3bf/caa3bd20-1bd2-11e7-b8ec-b883f2d63b7b.ico',
+      logo: '23f10080-faaa-11e6-b856-36051d50f3bf/757eb280-1df5-11e7-b8ec-b883f2d63b7b.png',
+      styles: {
+        productName: 'Mitel',
+        accentColor: '#FF7300',
+        accentHoverColor: '#E3F3FF',
+        navbar: '#15325F',
+        navbarText: '#FFFFFF',
+        primaryColor: '#00A1F4'
+      }
+    };
+
+    $rootScope.$on('$stateChangeStart', function(e, toState) {
+      if (toState.name === 'login' || toState.name === 'forgot-password') {
+        if ($location.absUrl().indexOf(mitelUrl) !== -1) {
+          Branding.set(mockBrandingData);
+        } else {
+          Branding.set({});
+        }
+      }
+    });
     $rootScope.$on('$stateChangeSuccess', function() {
       queryCache.removeAll();
-      $rootScope.title = $state.current.title + ' | CxEngage';
+      $rootScope.title = $state.current.title + ' | ' + $rootScope.productName;
     });
     $rootScope.$on('$stateChangeError', function() {
       console.error('State change error!');